Remove cooperage module and fold barrel planning into wikibase orchestration
